Factor out a service.rs file
